Renamed UnitsController and moved to Utils
